+++ ir/be/ia32/x86_asm.h
@@ -19,13 +19,15 @@
 #include "compiler.h"
 #include "firm_types.h"
 #include "util.h"
+#include "x86_address_mode.h"
 #include "x86_node.h"
 
 typedef struct x86_asm_operand_t {
 	be_asm_operand_t op;
 	union {
-		ir_mode    *mode;
-		x86_imm32_t imm32;
+		ir_mode     *mode;
+		x86_imm32_t  imm32;
+		x86_addr_t   addr;
 	} u;
 } x86_asm_operand_t;